+++ lib/src/components/inputs/tecta_radio.dart
@@ -78,7 +78,7 @@ class TectaRadio<T> extends StatelessWidget {
   bool get _isSelected => value == groupValue;
 
   Color _getColor(BuildContext context) {
-    final Color activeColor = color ?? TectaColors.primaryMain;
+    final Color activeColor = color ?? Theme.of(context).colorScheme.primary;
     return TectaColors.resolve(context, activeColor);
   }