+++ lib/src/components/inputs/tecta_switch.dart
@@ -72,7 +72,7 @@ class TectaSwitch extends StatelessWidget {
   });
 
   Color _getColor(BuildContext context) {
-    final Color activeColor = color ?? TectaColors.primaryMain;
+    final Color activeColor = color ?? Theme.of(context).colorScheme.primary;
     return TectaColors.resolve(context, activeColor);
   }